Statistic Frontend Außerdem Backend Änderungen die nötig wurden. DTOs DTOs DTOs...

+++ backend/src/main/java/org/example/edhtrack/dto/stats/CommanderAmountStatDTO.java
@@ -1,3 +1,3 @@
 package org.example.edhtrack.dto.stats;
 
-public record CommanderAmountStatDTO(String commander, long count) {}
+public record CommanderAmountStatDTO(String commander, int count) {}
